+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+<a name="1.1.2"></a>
+## 1.1.2 (2016-11-09)
+
+
+#### Bug Fixes
+
+*   fix more overflow errors ([3ed7004e](https://github.com/ccampo133/exponential-backoff/commit/3ed7004e0ce6001312978dfd368dc21920a0398d))
+*   fix bug where infinite and jitter flags were being mixed up ([11ca0b6c](https://github.com/ccampo133/exponential-backoff/commit/11ca0b6c3cd9555fb96a6ae77a70cc0574a3c8bb))
+*   fix typo in API (attemps -> attempts) ([58359cdc](https://github.com/ccampo133/exponential-backoff/commit/58359cdc4a36fe6d828e03be32356f317cc0b9da))
+
+
+
 <a name="1.1.1"></a>
 ## 1.1.1 (2016-11-02)